Vendoring in psych 0c82d6d009eceb2bac048c2262f68c1fc494dde9
Psych was having problems with aliases.  VCR uses Psych to serialize requests and aliases were being defined but not recognized when the fixtures were loaded.

There is a fix merged at this release, but it's not been released as a gem yet.  The gem has C Extensions and the repo doesn't have a built version of the gem, so vendoring is the easiest way to solve the problem.

NB:  Had to update to Rubygems 2.0 and Bundler 1.3.5 to allow psych to be loaded from the bundle's version, rather then loaded BY bundle from stdlib.