Exclude specs from build gem to speed install.
`gem.test-files` is not documented and a hypersearch of the digital
meganet shows that it maybe never did anything and probably doesn't
know.

All the data in `spec/fixtures` is _super_ slow to download so our tiny
widdle gem takes longer then almost anything else to fetch, for a
use-case many/most? users don't have: Running unit tests on install.

This change removes `gem.test-files` and also excludes the `spec`
directory from the gem when build.